# Conflicts:
#	src/Jam6/Assets/Scenes/Main.unity

+++ src/Jam6/Assets/Scripts/IDamageable.cs
@@ -1,5 +1,5 @@
 ﻿
 public interface IDamageable
 {
-    float ApplyDamage(float rawDamage);
+    int ApplyDamage(int rawDamage);
 }
